requirements.txt
asgiref==3.5.0
Django==4.0.4
paho-mqtt==1.6.1
python-dotenv==0.20.0
sqlparse==0.4.2
